fix: sqrtf -> sqrt

+++ analysis/tests/simd/vec/AVX/avx_float_vec2_tests.cpp
@@ -264,7 +264,7 @@ namespace tests
 
         for (size_t i = 0; i < vecLen; i++)
         {
-            float expected = std::sqrtf(vecData[i]);
+            float expected = std::sqrt(vecData[i]);
             EXPECT_NEAR(res.data[i], expected, maximalFastInvSqrtRelativeError * std::abs(expected == 0 ? 0.1 : expected));
         }
     }
@@ -279,7 +279,7 @@ namespace tests
 
         for (size_t i = 0; i < vecLen; i++)
         {
-            float expected = 1 / std::sqrtf(vecData[i]);
+            float expected = 1 / std::sqrt(vecData[i]);
             EXPECT_NEAR(res.data[i], expected, maximalFastInvSqrtRelativeError * std::abs(expected == 0 ? 0.1 : expected));
         }
     }
