+++ analysis/tests/simd/vec/AVX/avx_float_vec3_tests.cpp
@@ -266,7 +266,7 @@ namespace tests
 
         for (size_t i = 0; i < vecLen; i++)
         {
-            float expected = std::sqrtf(vecData[i]);
+            float expected = std::sqrt(vecData[i]);
             EXPECT_NEAR(res.data[i], expected, maximalFastInvSqrtRelativeError * std::abs(expected == 0 ? 0.1 : expected));
         }
     }
@@ -281,7 +281,7 @@ namespace tests
 
         for (size_t i = 0; i < vecLen; i++)
         {
-            float expected = 1 / std::sqrtf(vecData[i]);
+            float expected = 1 / std::sqrt(vecData[i]);
             EXPECT_NEAR(res.data[i], expected, maximalFastInvSqrtRelativeError * std::abs(expected == 0 ? 0.1 : expected));
         }
     }
